requirements.txt
# ML Libraries
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2
